usize -> u32 in QueryData
We don't expect more than 2^32 queries in a given execution and there
are a *lot* of these structs in memory, so we need to try to keep its
size down.

+++ site/src/api.rs
@@ -264,9 +264,9 @@ pub mod self_profile {
         pub label: String,
         pub self_time: Duration,
         pub percent_total_time: f32,
-        pub number_of_cache_misses: usize,
-        pub number_of_cache_hits: usize,
-        pub invocation_count: usize,
+        pub number_of_cache_misses: u32,
+        pub number_of_cache_hits: u32,
+        pub invocation_count: u32,
         pub blocked_time: Duration,
         pub incremental_load_time: Duration,
     }
